Deployment/requirements.txt
streamlit
pandas
numpy
seaborn
matplotlib
Pillow
plotly
scikit-learn == 1.2.2
nltk==3.8.1
wordcloud==1.9.3
tensorflow==2.15.0
tensorflow_hub==0.16.1